test: fold join link tests into their feature files per the test policy
CLAUDE.md and CONTRIBUTING.md (Tests): one file per feature area, never a new
test file per change, and no test that checks copy or restates a helper.

- e2e/join-link.smoke.spec.ts moves into e2e/smoke.spec.ts, unchanged.
- src/pages/settings/join-link-copy.test.ts is deleted outright: every case
  asserted a copy string or restated `price()`. The e2e smoke test exercises the
  card, and billing-plans.test.ts already pins unitPrice.

The api half moved in the previous commit, where the new pending-invite case
needed it.
